re order tests
