+++ frontend/src/pages/Admin/TrainingReports.js
@@ -6,23 +6,6 @@ function TrainingReports() {
   const typeName = 'training reports';
   const apiPathName = 'training-reports';
 
-  /*
-  const validCsvHeaders = [
-    'Sheet Name',
-    'Event ID',
-    'Edit Title',
-    'IST Name:',
-    'Creator',
-    'Event Organizer - Type of Event',
-    'National Center(s) Requested',
-    'Event Duration/# NC Days of Support',
-    'Reason for Activity',
-    'Target Population(s)',
-    'Audience',
-    'Overall Vision/Goal for the PD Event',
-  ];
-  */
-
   const requiredCsvHeaders = [
     'Event ID',
     'Edit Title',
